+++ backbone/src/main/java/org/researchstack/backbone/ui/ViewTaskActivity.java
@@ -3,14 +3,24 @@ package org.researchstack.backbone.ui;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
+import android.graphics.Color;
+import android.graphics.drawable.ColorDrawable;
+import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
+import android.view.Window;
+import android.view.WindowManager;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.Toast;
+
+import com.afollestad.materialdialogs.MaterialDialog;
+import com.afollestad.materialdialogs.Theme;
 
 import org.researchstack.backbone.R;
 import org.researchstack.backbone.result.StepResult;
@@ -18,7 +28,9 @@ import org.researchstack.backbone.result.TaskResult;
 import org.researchstack.backbone.step.Step;
 import org.researchstack.backbone.task.Task;
 import org.researchstack.backbone.ui.callbacks.StepCallbacks;
+import org.researchstack.backbone.ui.step.layout.ConsentVisualStepLayout;
 import org.researchstack.backbone.ui.step.layout.StepLayout;
+import org.researchstack.backbone.ui.step.layout.SurveyStepLayout;
 import org.researchstack.backbone.ui.views.StepSwitcher;
 
 import java.lang.reflect.Constructor;
@@ -28,16 +40,44 @@ public class ViewTaskActivity extends PinCodeActivity implements StepCallbacks {
     public static final String EXTRA_TASK = "ViewTaskActivity.ExtraTask";
     public static final String EXTRA_TASK_RESULT = "ViewTaskActivity.ExtraTaskResult";
     public static final String EXTRA_STEP = "ViewTaskActivity.ExtraStep";
+    public static final String EXTRA_COLOR_PRIMARY = "ViewTaskActivity.ExtraColorPrimary";
+    public static final String EXTRA_COLOR_PRIMARY_DARK = "ViewTaskActivity.ExtraColorPrimaryDark";
+    public static final String EXTRA_COLOR_SECONDARY = "ViewTaskActivity.ExtraColorSecondary";
+    public static final String EXTRA_PRINCIPAL_TEXT_COLOR = "ViewTaskActivity.ExtraPrincipalTextColor";
+    public static final String EXTRA_SECONDARY_TEXT_COLOR = "ViewTaskActivity.ExtraSecondaryTextColor";
+    public static final String EXTRA_ACTION_FAILED_COLOR = "ViewTaskActivity.ExtraActionFailedColor";
+    public static final String EXTRA_SHOW_CANCEL = "ViewTaskActivity.ExtraShowCancel";
 
     private StepSwitcher root;
 
     private Step currentStep;
     private Task task;
     private TaskResult taskResult;
+    private int colorPrimary;
+    private int colorPrimaryDark;
+    private int colorSecondary;
+    private int principalTextColor;
+    private int secondaryTextColor;
+    private int actionFailedColor;
+    private boolean showCancel;
 
     public static Intent newIntent(Context context, Task task) {
         Intent intent = new Intent(context, ViewTaskActivity.class);
         intent.putExtra(EXTRA_TASK, task);
+        return intent;
+    }
+
+    public static Intent newThemedIntent(Context context, Task task, boolean showCancel, int colorPrimary, int colorPrimaryDark, int colorSecondary,
+                                         int principalTextColor, int secondaryTextColor, int actionFailedColor) {
+        Intent intent = new Intent(context, ViewTaskActivity.class);
+        intent.putExtra(EXTRA_TASK, task);
+        intent.putExtra(EXTRA_COLOR_PRIMARY, colorPrimary);
+        intent.putExtra(EXTRA_COLOR_PRIMARY_DARK, colorPrimaryDark);
+        intent.putExtra(EXTRA_COLOR_SECONDARY, colorSecondary);
+        intent.putExtra(EXTRA_PRINCIPAL_TEXT_COLOR, principalTextColor);
+        intent.putExtra(EXTRA_SECONDARY_TEXT_COLOR, secondaryTextColor);
+        intent.putExtra(EXTRA_ACTION_FAILED_COLOR, actionFailedColor);
+        intent.putExtra(EXTRA_SHOW_CANCEL, showCancel);
         return intent;
     }
 
@@ -55,6 +95,13 @@ public class ViewTaskActivity extends PinCodeActivity implements StepCallbacks {
 
         if (savedInstanceState == null) {
             task = (Task) getIntent().getSerializableExtra(EXTRA_TASK);
+            colorPrimary = getIntent().getIntExtra(EXTRA_COLOR_PRIMARY, R.color.rsb_colorPrimary);
+            colorPrimaryDark = getIntent().getIntExtra(EXTRA_COLOR_PRIMARY_DARK, R.color.rsb_colorPrimaryDark);
+            colorSecondary = getIntent().getIntExtra(EXTRA_COLOR_SECONDARY, R.color.rsb_colorAccent);
+            principalTextColor = getIntent().getIntExtra(EXTRA_PRINCIPAL_TEXT_COLOR, R.color.rsb_cell_header_grey);
+            secondaryTextColor = getIntent().getIntExtra(EXTRA_SECONDARY_TEXT_COLOR, R.color.rsb_item_text_grey);
+            actionFailedColor = getIntent().getIntExtra(EXTRA_ACTION_FAILED_COLOR, R.color.rsb_error);
+            showCancel = getIntent().getBooleanExtra(EXTRA_SHOW_CANCEL, false);
             taskResult = new TaskResult(task.getIdentifier());
             taskResult.setStartDate(new Date());
         } else {
@@ -96,8 +143,7 @@ public class ViewTaskActivity extends PinCodeActivity implements StepCallbacks {
     }
 
     private void showStep(Step step) {
-        int currentStepPosition = task.getProgressOfCurrentStep(currentStep, taskResult)
-                .getCurrent();
+        int currentStepPosition = task.getProgressOfCurrentStep(currentStep, taskResult).getCurrent();
         int newStepPosition = task.getProgressOfCurrentStep(step, taskResult).getCurrent();
 
         StepLayout stepLayout = getLayoutForStep(step);
@@ -113,13 +159,22 @@ public class ViewTaskActivity extends PinCodeActivity implements StepCallbacks {
         // Change the title on the activity
         String title = task.getTitleForStep(this, step);
         setActionBarTitle(title);
+        step.setStepTheme(colorPrimary, colorPrimaryDark, colorSecondary, principalTextColor, secondaryTextColor, actionFailedColor);
+        setActivityTheme(colorPrimary, colorPrimaryDark);
 
         // Get result from the TaskResult, can be null
         StepResult result = taskResult.getStepResult(step.getIdentifier());
 
         // Return the Class & constructor
         StepLayout stepLayout = createLayoutFromStep(step);
-        stepLayout.initialize(step, result);
+        if (stepLayout instanceof SurveyStepLayout) {
+            ((SurveyStepLayout) stepLayout).initialize(step, result, colorPrimary, colorSecondary, principalTextColor, secondaryTextColor);
+        } else if (stepLayout instanceof ConsentVisualStepLayout) {
+            ((ConsentVisualStepLayout) stepLayout).initialize(step, result, colorPrimary, colorSecondary, principalTextColor, secondaryTextColor);
+        } else {
+            stepLayout.initialize(step, result);
+        }
+
         stepLayout.setCallbacks(this);
 
         return stepLayout;
@@ -165,10 +220,29 @@ public class ViewTaskActivity extends PinCodeActivity implements StepCallbacks {
     }
 
     @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.rsb_activity_view_task_menu, menu);
+        menu.findItem(R.id.rsb_action_cancel).setVisible(showCancel);
+        return true;
+    }
+
+    @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         if (item.getItemId() == android.R.id.home) {
             notifyStepOfBackPress();
             return true;
+        } else if (item.getItemId() == R.id.rsb_action_cancel) {
+            new MaterialDialog.Builder(this)
+                    .title(R.string.rsb_task_cancel_title)
+                    .content(R.string.rsb_task_cancel_text)
+                    .theme(Theme.LIGHT)
+                    .positiveColor(colorPrimary)
+                    .negativeColor(colorPrimary)
+                    .negativeText(R.string.rsb_cancel)
+                    .positiveText(R.string.rsb_task_cancel_positive)
+                    .onPositive((dialog, which) -> finish())
+                    .show();
         }
 
         return super.onOptionsItemSelected(item);
@@ -264,5 +338,24 @@ public class ViewTaskActivity extends PinCodeActivity implements StepCallbacks {
         if (actionBar != null) {
             actionBar.setTitle(title);
         }
+    }
+
+    private void setActivityTheme(int primaryColor, int primaryColorDark) {
+        runOnUiThread(() -> {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                Window window = getWindow();
+
+                if (primaryColorDark == Color.BLACK && window.getNavigationBarColor() == Color.BLACK) {
+                    window.clearFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
+                } else {
+                    window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
+                }
+                window.setStatusBarColor(primaryColorDark);
+            }
+            ActionBar actionBar = getSupportActionBar();
+            if (actionBar != null) {
+                actionBar.setBackgroundDrawable(new ColorDrawable(primaryColor));
+            }
+        });
     }
 }