+++ backbone/src/main/java/org/researchstack/backbone/ui/step/body/BodyAnswer.java
@@ -7,8 +7,7 @@ import org.researchstack.backbone.R;
 
 public class BodyAnswer {
     public static final BodyAnswer VALID = new BodyAnswer(true, 0);
-    public static final BodyAnswer INVALID = new BodyAnswer(false,
-            R.string.rsb_invalid_answer_default);
+    public static final BodyAnswer INVALID = new BodyAnswer(false, R.string.rsb_invalid_answer_default);
 
     private boolean isValid;
     private int reason;
@@ -18,6 +17,12 @@ public class BodyAnswer {
         this.isValid = isValid;
         this.reason = reason;
         this.params = params;
+    }
+
+    public BodyAnswer(boolean isValid, @StringRes int reason) {
+        this.isValid = isValid;
+        this.reason = reason;
+        this.params = new String[0];
     }
 
     public boolean isValid() {