+++ backbone/src/main/java/org/researchstack/backbone/ui/step/layout/ConsentSignatureStepLayout.java
@@ -1,11 +1,15 @@
 package org.researchstack.backbone.ui.step.layout;
 
+import android.app.Activity;
 import android.content.Context;
 import android.graphics.Bitmap;
+import android.support.v7.widget.AppCompatTextView;
 import android.util.AttributeSet;
 import android.util.Base64;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
+import android.view.inputmethod.InputMethodManager;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -36,23 +40,29 @@ public class ConsentSignatureStepLayout extends RelativeLayout implements StepLa
     private StepCallbacks callbacks;
     private Step step;
     private StepResult<String> result;
+    private Context context;
 
     public ConsentSignatureStepLayout(Context context) {
         super(context);
+        this.context = context;
     }
 
     public ConsentSignatureStepLayout(Context context, AttributeSet attrs) {
         super(context, attrs);
+        this.context = context;
     }
 
     public ConsentSignatureStepLayout(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
+        this.context = context;
     }
 
     @Override
     public void initialize(Step step, StepResult result) {
         this.step = step;
         this.result = result == null ? new StepResult<>(step) : result;
+
+        hideSoftKeyboard();
 
         initializeStep();
     }
@@ -75,16 +85,17 @@ public class ConsentSignatureStepLayout extends RelativeLayout implements StepLa
     }
 
     private void initializeStep() {
-        LayoutInflater.from(getContext())
-                .inflate(R.layout.rsb_step_layout_consent_signature, this, true);
+        LayoutInflater.from(getContext()).inflate(R.layout.rsb_step_layout_consent_signature, this, true);
 
         TextView title = (TextView) findViewById(R.id.title);
+        title.setTextColor(step.getPrincipalTextColor());
         title.setText(step.getTitle());
 
         TextView text = (TextView) findViewById(R.id.summary);
         text.setText(step.getText());
 
-        View clear = findViewById(R.id.layout_consent_review_signature_clear);
+        AppCompatTextView clear = (AppCompatTextView) findViewById(R.id.layout_consent_review_signature_clear);
+        clear.setTextColor(step.getPrimaryColor());
 
         signatureView = (SignatureView) findViewById(R.id.layout_consent_review_signature);
         signatureView.setCallbacks(new SignatureCallbacks() {
@@ -112,6 +123,8 @@ public class ConsentSignatureStepLayout extends RelativeLayout implements StepLa
 
         SubmitBar submitBar = (SubmitBar) findViewById(R.id.submit_bar);
         submitBar.getNegativeActionView().setVisibility(View.GONE);
+        submitBar.setPositiveTitleColor(step.getColorSecondary());
+        submitBar.setPositiveTitle(R.string.rsb_done);
         submitBar.setPositiveAction(v -> {
             if (signatureView.isSignatureDrawn()) {
                 setDataToResult();
@@ -146,4 +159,13 @@ public class ConsentSignatureStepLayout extends RelativeLayout implements StepLa
         }
     }
 
+    private void hideSoftKeyboard() {
+        try {
+            InputMethodManager imm = (InputMethodManager) context.getSystemService(Activity.INPUT_METHOD_SERVICE);
+            imm.hideSoftInputFromWindow(getRootView().getWindowToken(), 0);
+            getRootView().clearFocus();
+        } catch (NullPointerException e) {
+            Log.e("CSSL", "NPE: " + e.getMessage());
+        }
+    }
 }