+++ backbone/src/main/java/org/researchstack/backbone/ui/step/layout/ConsentVisualStepLayout.java
@@ -2,16 +2,12 @@ package org.researchstack.backbone.ui.step.layout;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.res.TypedArray;
 import android.graphics.drawable.Drawable;
 import android.support.v4.graphics.drawable.DrawableCompat;
 import android.util.AttributeSet;
-import android.util.TypedValue;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
-
-import com.jakewharton.rxbinding.view.RxView;
 
 import org.researchstack.backbone.R;
 import org.researchstack.backbone.model.ConsentSection;
@@ -25,10 +21,16 @@ import org.researchstack.backbone.ui.views.SubmitBar;
 import org.researchstack.backbone.utils.ResUtils;
 import org.researchstack.backbone.utils.TextUtils;
 
+import java.net.URL;
+
 public class ConsentVisualStepLayout extends FixedSubmitBarLayout implements StepLayout {
 
     private StepCallbacks callbacks;
     private ConsentVisualStep step;
+    private int colorPrimary;
+    private int colorSecondary;
+    private int principalTextColor;
+    private int secondaryTextColor;
 
     public ConsentVisualStepLayout(Context context) {
         super(context);
@@ -45,6 +47,15 @@ public class ConsentVisualStepLayout extends FixedSubmitBarLayout implements Ste
     @Override
     public void initialize(Step step, StepResult result) {
         this.step = (ConsentVisualStep) step;
+        initializeStep();
+    }
+
+    public void initialize(Step step, StepResult result, int colorPrimary, int colorSecondary, int principalTextColor, int secondaryTextColor) {
+        this.step = (ConsentVisualStep) step;
+        this.colorPrimary = colorPrimary;
+        this.colorSecondary = colorSecondary;
+        this.principalTextColor = principalTextColor;
+        this.secondaryTextColor = secondaryTextColor;
         initializeStep();
     }
 
@@ -70,14 +81,7 @@ public class ConsentVisualStepLayout extends FixedSubmitBarLayout implements Ste
     }
 
     private void initializeStep() {
-        ConsentSection data = step.getSection();
-
-        // Set Image
-        TypedValue typedValue = new TypedValue();
-        TypedArray a = getContext().obtainStyledAttributes(typedValue.data,
-                new int[]{R.attr.colorAccent});
-        int accentColor = a.getColor(0, 0);
-        a.recycle();
+        final ConsentSection data = step.getSection();
 
         ImageView imageView = (ImageView) findViewById(R.id.image);
 
@@ -87,13 +91,25 @@ public class ConsentVisualStepLayout extends FixedSubmitBarLayout implements Ste
 
         int imageResId = ResUtils.getDrawableResourceId(getContext(), imageName);
 
+        Drawable drawable = null;
+
         if (imageResId != 0) {
-            Drawable drawable = getResources().getDrawable(imageResId);
+            drawable = getResources().getDrawable(imageResId);
+        }
+        else if(imageName != null && imageResId == 0)
+        {
+            drawable = Drawable.createFromPath(imageName);
+        }
+
+        if (drawable != null)
+        {
             drawable = DrawableCompat.wrap(drawable);
-            DrawableCompat.setTint(drawable, accentColor);
+            DrawableCompat.setTint(drawable, colorSecondary);
             imageView.setImageDrawable(drawable);
             imageView.setVisibility(View.VISIBLE);
-        } else {
+        }
+        else
+        {
             imageView.setVisibility(View.GONE);
         }
 
@@ -102,6 +118,7 @@ public class ConsentVisualStepLayout extends FixedSubmitBarLayout implements Ste
         String title = TextUtils.isEmpty(data.getTitle()) ? getResources().getString(data.getType()
                 .getTitleResId()) : data.getTitle();
         titleView.setText(title);
+        titleView.setTextColor(principalTextColor);
 
         // Set Summary
         TextView summaryView = (TextView) findViewById(R.id.summary);
@@ -117,21 +134,38 @@ public class ConsentVisualStepLayout extends FixedSubmitBarLayout implements Ste
                 moreInfoView.setText(data.getType().getMoreInfoResId());
             }
 
-            RxView.clicks(moreInfoView).subscribe(v -> {
-                String webTitle = getResources().getString(R.string.rsb_consent_section_more_info);
-                Intent webDoc = ViewWebDocumentActivity.newIntentForContent(getContext(), webTitle,
-                        TextUtils.isEmpty(data.getContent()) ? data.getHtmlContent() : data.getContent());
-                getContext().startActivity(webDoc);
+            moreInfoView.setOnClickListener(new OnClickListener() {
+
+                @Override
+                public void onClick(View view) {
+                    String webTitle = getResources().getString(R.string.rsb_consent_section_more_info);
+                    URL contentUrl = data.getContentUrl();
+                    Intent webDoc;
+                    if(contentUrl != null)
+                    {
+                         webDoc = ViewWebDocumentActivity.newIntentForContent(getContext(), webTitle, contentUrl, true);
+
+                    }
+                    else
+                    {
+                        webDoc = ViewWebDocumentActivity.newIntentForContent(getContext(), webTitle,
+                                TextUtils.isEmpty(data.getContent()) ? data.getHtmlContent() : data.getContent());
+                    }
+
+                    ViewWebDocumentActivity.addThemeColors(webDoc, step.getPrimaryColor(), step.getColorPrimaryDark());
+                    getContext().startActivity(webDoc);
+                }
             });
+
         } else {
             moreInfoView.setVisibility(View.GONE);
         }
 
         SubmitBar submitBar = (SubmitBar) findViewById(R.id.rsb_submit_bar);
         submitBar.setPositiveTitle(step.getNextButtonString());
-        submitBar.setPositiveAction(v -> callbacks.onSaveStep(StepCallbacks.ACTION_NEXT,
-                step,
-                null));
+        submitBar.setNegativeTitleColor(colorPrimary);
+        submitBar.setPositiveTitleColor(colorSecondary);
+        submitBar.setPositiveAction(v -> callbacks.onSaveStep(StepCallbacks.ACTION_NEXT, step, null));
         submitBar.getNegativeActionView().setVisibility(View.GONE);
     }
 }