+++ backbone/src/main/java/org/researchstack/backbone/ui/step/layout/SurveyStepLayout.java
@@ -49,6 +49,11 @@ public class SurveyStepLayout extends FixedSubmitBarLayout implements StepLayout
     private LinearLayout container;
     private StepBody stepBody;
 
+    private int coloryPrimary;
+    private int colorSecondary;
+    private int principalTextColor;
+    private int secondaryTextColor;
+
     public SurveyStepLayout(Context context) {
         super(context);
     }
@@ -70,6 +75,22 @@ public class SurveyStepLayout extends FixedSubmitBarLayout implements StepLayout
         if (!(step instanceof QuestionStep)) {
             throw new RuntimeException("Step being used in SurveyStep is not a QuestionStep");
         }
+
+        this.questionStep = (QuestionStep) step;
+        this.stepResult = result;
+
+        initializeStep();
+    }
+
+    public void initialize(Step step, StepResult result, int colorPrimary, int colorSecondary, int principalTextColor, int secondaryTextColor) {
+        if (!(step instanceof QuestionStep)) {
+            throw new RuntimeException("Step being used in SurveyStep is not a QuestionStep");
+        }
+
+        this.coloryPrimary = colorPrimary;
+        this.colorSecondary = colorSecondary;
+        this.principalTextColor = principalTextColor;
+        this.secondaryTextColor = secondaryTextColor;
 
         this.questionStep = (QuestionStep) step;
         this.stepResult = result;
@@ -113,8 +134,12 @@ public class SurveyStepLayout extends FixedSubmitBarLayout implements StepLayout
 
         container = (LinearLayout) findViewById(R.id.rsb_survey_content_container);
         TextView title = (TextView) findViewById(R.id.rsb_survey_title);
+        title.setTextColor(principalTextColor);
         TextView summary = (TextView) findViewById(R.id.rsb_survey_text);
+        summary.setTextColor(secondaryTextColor);
         SubmitBar submitBar = (SubmitBar) findViewById(R.id.rsb_submit_bar);
+        submitBar.setNegativeTitleColor(coloryPrimary);
+        submitBar.setPositiveTitleColor(colorSecondary);
         submitBar.setPositiveAction(v -> onNextClicked());
 
         if (questionStep != null) {
@@ -213,5 +238,4 @@ public class SurveyStepLayout extends FixedSubmitBarLayout implements StepLayout
     public String getString(@StringRes int stringResId) {
         return getResources().getString(stringResId);
     }
-
 }