+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ### Change log
 
+#### v1.0.3
+
+- Added the ability to theme RS.
+- Added support for custom consent sections:
+   - Formal title.
+   - Html Content.
+   - Custom Url content.
+   - Custom image.
+- Bug fixes & improvements.
+
+#### v1.0.2
+
+- Bug fixes & improvements.
+
 #### v1.0.1
 
 - Task Activities saves the state on when they finish with unsuccess.